Completed Backend V2 (With Spring Security with JWT based AUTH)

+++ Nrv-Blog-API/src/main/java/com/nrv/NrvBlogAPI/service/UserServiceImpl.java
@@ -1,5 +1,6 @@
 package com.nrv.NrvBlogAPI.service;
 
+import com.nrv.NrvBlogAPI.custom_exception.InvalidCredentialsException;
 import com.nrv.NrvBlogAPI.custom_exception.ResourceNotFoundException;
 import com.nrv.NrvBlogAPI.dto.APIResponse;
 import com.nrv.NrvBlogAPI.dto.userDTO.*;
@@ -8,11 +9,18 @@ import com.nrv.NrvBlogAPI.entities.User;
 import com.nrv.NrvBlogAPI.log.UserLogMessages;
 import com.nrv.NrvBlogAPI.repository.BlogRepository;
 import com.nrv.NrvBlogAPI.repository.UserRepository;
+import com.nrv.NrvBlogAPI.security.JwtUtils;
 import jakarta.transaction.Transactional;
 import org.modelmapper.ModelMapper;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
@@ -34,15 +42,18 @@ public class UserServiceImpl implements UserService {
     @Autowired
     ModelMapper mapper;
 
-/*    @Autowired
-    private PasswordEncoder encoder;*/
+    @Autowired
+    private PasswordEncoder encoder;
 
-/*    @Autowired
-    private AuthenticationManager mgr;*/
+    @Autowired
+    private AuthenticationManager mgr;
+
+    @Autowired
+    private JwtUtils utils;
 
     @Override
     public UserDTO getUserWithBlogs(String userId) {
-        logger.trace(UserLogMessages.USER_GET.getMessage(), userId);
+        logger.info(UserLogMessages.USER_GET.getMessage(), userId);
         User user = userRepository
                 .findByUserIdWithBlogs(userId)
                 .orElseThrow(() -> new ResourceNotFoundException("User with " + userId + " not found"));
@@ -61,29 +72,26 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public LoginResponseDTO userLogin(LoginUserDTO user) {
-/*
         String jwtToken;
         try {
-            // Perform authentication
-            Authentication principal = mgr.authenticate(
-                    new UsernamePasswordAuthenticationToken(user.getEmail(), user.getPassword()));
+            Authentication principal = mgr
+                    .authenticate(
+                            new UsernamePasswordAuthenticationToken(
+                                    user.getUserId(),
+                                    user.getPassword()
+                            ));
 
             // Generate JWT token if authentication is successful
             jwtToken = utils.generateJwtToken(principal);
 
-            // Log the successful login
-            logger.trace(UserLogMessages.USER_LOGIN.getMessage(), user.getUserId());
+            logger.info(UserLogMessages.USER_LOGIN.getMessage(), user.getUserId());
 
         } catch (AuthenticationException e) {
             // Handle authentication failure
-            logger.error("Authentication failed for user: {}", user.getEmail(), e);
+            logger.error("Authentication failed for user: {}", user.getUserId(), e);
             throw new InvalidCredentialsException("Invalid email or password.");
         }
-
-        // Return successful response
         return new LoginResponseDTO(user.getUserId(), jwtToken, "User Logged In Successfully");
-*/
-        return null;
     }
 
     @Override
@@ -92,34 +100,42 @@ public class UserServiceImpl implements UserService {
                 userRegistration.getUserId(),
                 userRegistration.getUserName(),
                 Role.ROLE_USER, // There's only One admin so by default any other are ROLE_USER
-                //encoder.encode(
-                userRegistration.getPassword(),
-                //),
+                encoder.encode(
+                        userRegistration.getPassword()
+                ),
                 LocalDate.now(),
                 List.of()
         );
-
+        userRepository.save(newUser);
+        logger.info(UserLogMessages.USER_REGISTER.getMessage(), newUser.getUserId());
         return mapper.map(newUser, UserRegisteredDTO.class);
     }
 
     @Override
     public APIResponse userUpdate(UpdateUserDTO updatedUser) {
-        User user = userRepository
-                .findById(updatedUser.getUserId())
-                .orElseThrow(() -> new ResourceNotFoundException("User with That ID Not found with that ID"));
-/*        if (!encoder.matches(updatedUser.getPassword(), user.getPassword())) {
-            throw new InvalidPasswordException("Old password does not match");
+        //Implement logic of the that user can only update their info.
+        String currentUsername = SecurityContextHolder.getContext().getAuthentication().getName();
+        if (currentUsername.equals(updatedUser.getUserId())) {
+            User user = userRepository
+                    .findById(updatedUser.getUserId())
+                    .orElseThrow(() -> new ResourceNotFoundException("User with That ID Not found with that ID"));
+            if (encoder.matches(updatedUser.getPassword(), user.getPassword())) {
+                throw new InvalidCredentialsException("Old password and New Password Are Same");
+            }
+            user.setPassword(encoder.encode(updatedUser.getPassword()));
+            userRepository.save(user);
+            logger.info(UserLogMessages.USER_UPDATE.getMessage(), updatedUser.getUserId());
+            return new APIResponse("User Password Updated");
+        } else {
+            return new APIResponse("You can't update the user");
         }
-        user.setPassword(encoder.encode(updatedUser.getPassword()));*/
-        userRepository.save(user);
-        return new APIResponse("User Password Updated");
     }
 
     @Override
     public List<UserDTO> getListOfUsers() {
         // Fetch all users with their associated blogs in one query
         List<User> userList = userRepository.findAllWithBlogs();
-        logger.trace(UserLogMessages.USER_GET_LIST.getMessage());
+        logger.info(UserLogMessages.USER_GET_LIST.getMessage());
         // Convert the list of users to a list of UserDTOs
         List<UserDTO> userDTOList = new ArrayList<>();
         for (User user : userList) {
@@ -139,8 +155,21 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public DeleteUserResponseDto deleteAUser(String userId) {
-        userRepository.deleteById(userId);
-        logger.trace(UserLogMessages.BLOG_DELETED.getMessage(), userId);
+        //Logic of Admin or User only
+        String currentUsername = SecurityContextHolder.getContext().getAuthentication().getName();
+        User deleteUser = userRepository.findById(userId)
+                .orElseThrow(()-> new ResourceNotFoundException("User with this ID not present"));
+        if(deleteUser.getUserId().equals(currentUsername) || !isAdmin()){
+            throw new RuntimeException("You are not authorized to delete this user");
+        }
+        userRepository.delete(deleteUser);
+        logger.info(UserLogMessages.BLOG_DELETED.getMessage(), userId);
         return new DeleteUserResponseDto("User with ID: " + userId + " Deleted");
+    }
+
+    private boolean isAdmin() {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        return auth.getAuthorities().stream()
+                .anyMatch(grantedAuthority -> grantedAuthority.getAuthority().equals("ROLE_ADMIN"));
     }
 }
